Use portable format specifiers: Case %ld - localized - smsg.
Fix uses of localized "%ld" within smsg():
- Replace "%ld" with "%" PRId64.
- Cast corresponding argument to (int64_t).

+++ src/ops.c
@@ -594,7 +594,7 @@ int         (*how)(void);
     if (i > 1
         && (i % 50 == 0 || i == oap->line_count - 1)
         && oap->line_count > p_report)
-      smsg((char_u *)_("%ld lines to indent... "), i);
+      smsg((char_u *)_("%" PRId64 " lines to indent... "), (int64_t)i);
 
     /*
      * Be vi-compatible: For lisp indenting the first line is not
@@ -638,7 +638,7 @@ int         (*how)(void);
     if (i == 1)
       MSG(_("1 line indented "));
     else
-      smsg((char_u *)_("%ld lines indented "), i);
+      smsg((char_u *)_("%" PRId64 " lines indented "), (int64_t)i);
   }
   /* set '[ and '] marks */
   curbuf->b_op_start = oap->start;
@@ -1953,7 +1953,7 @@ void op_tilde(oparg_T *oap)
     if (oap->line_count == 1)
       MSG(_("1 line changed"));
     else
-      smsg((char_u *)_("%ld lines changed"), oap->line_count);
+      smsg((char_u *)_("%" PRId64 " lines changed"), (int64_t)oap->line_count);
   }
 }
 
@@ -2556,9 +2556,10 @@ int op_yank(oparg_T *oap, int deleting, int mess)
         else
           MSG(_("1 line yanked"));
       } else if (oap->block_mode)
-        smsg((char_u *)_("block of %ld lines yanked"), yanklines);
+        smsg((char_u *)_("block of %" PRId64 " lines yanked"),
+             (int64_t)yanklines);
       else
-        smsg((char_u *)_("%ld lines yanked"), yanklines);
+        smsg((char_u *)_("%" PRId64 " lines yanked"), (int64_t)yanklines);
     }
   }
